refactor: move ResourceIndicator interface to web-pkg
